pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "anthropic>=0.49.0",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.3.1",
    "moviepy>=2.1.2",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "opencv-python>=4.11.0.86",
    "pydub>=0.25.1",
    "streamlit>=1.44.1",
    "trafilatura>=2.0.0",
]
